First iteration of `AddSubscriptionScreen` cleanup
